Added RK8 as time integrate for Burgers.
Not yet added to the solution. Ideally we'd have a combined
implementation, but that implementation needs to handle both situations of
stopping for agent actions and processing the whole thing at once.
I don't think I've quite implemented this correctly.

Also some tweaks to the convergence plot scripts.

+++ envs/burgers_env.py
@@ -573,7 +573,7 @@ class SplitFluxBurgersEnv(AbstractBurgersEnv, Plottable1DEnv):
 
         self.action_space = SoftmaxBox(low=-np.inf, high=np.inf,
                                        shape=(self.grid.real_length() + 1, 2, 2 * self.weno_order - 1),
-                                       dtype=np.float32)
+                                       dtype=np.float64)
         self.observation_space = spaces.Box(low=-1e7, high=1e7,
                                             shape=(self.grid.real_length() + 1, 2, 2 * self.state_order - 1),
                                             dtype=np.float64)
@@ -653,7 +653,7 @@ class FluxBurgersEnv(AbstractBurgersEnv):
 
         self.action_space = SoftmaxBox(low=-np.inf, high=np.inf,
                                        shape=(self.grid.real_length() + 1, 2 * self.weno_order),
-                                       dtype=np.float32)
+                                       dtype=np.float64)
         self.observation_space = spaces.Box(low=-1e7, high=1e7,
                                             shape=(self.grid.real_length() + 1, 2 * self.state_order),
                                             dtype=np.float64)
